many thing

+++ src/uet/oop/bomberman/entities/powers/PowerUp.java
@@ -5,7 +5,7 @@ import uet.oop.bomberman.entities.Entity;
 
 
 public abstract class PowerUp extends Entity {
-    public PowerUp(int xUnit, int yUnit, Image image){
+    public PowerUp(double xUnit, double yUnit, Image image){
         super(xUnit, yUnit, image);
     }
 }
