+++ src/uet/oop/bomberman/entities/powers/PowerUpBombs.java
@@ -3,7 +3,7 @@ package uet.oop.bomberman.entities.powers;
 import javafx.scene.image.Image;
 
 public class PowerUpBombs extends PowerUp{
-    public PowerUpBombs(int xUnit, int yUnit, Image image){
+    public PowerUpBombs(double xUnit, double yUnit, Image image){
         super(xUnit, yUnit, image);
     }
     @Override
