+++ src/uet/oop/bomberman/entities/powers/PowerUpFlames.java
@@ -3,7 +3,7 @@ package uet.oop.bomberman.entities.powers;
 import javafx.scene.image.Image;
 
 public class PowerUpFlames extends PowerUp{
-    public PowerUpFlames(int xUnit, int yUnit, Image image){
+    public PowerUpFlames(double xUnit, double yUnit, Image image){
         super(xUnit, yUnit, image);
     }
