+++ src/uet/oop/bomberman/entities/powers/PowerUpSpeed.java
@@ -3,7 +3,7 @@ package uet.oop.bomberman.entities.powers;
 import javafx.scene.image.Image;
 
 public class PowerUpSpeed extends PowerUp{
-    public PowerUpSpeed(int xUnit, int yUnit, Image image){
+    public PowerUpSpeed(double xUnit, double yUnit, Image image){
         super(xUnit, yUnit, image);
     }
     @Override
